assets → copy, remove imagemin dependency
See: #26 cloudfour/drizzle#33
