+++ test/fixtures/static/favicon.ico
