+++ test/gulpfile.js
@@ -3,13 +3,13 @@
 var gulp = require('gulp');
 var tasks = require('../');
 
-tasks.assets(gulp, {
-  src: './fixtures/assets/*',
-  dest: './dist'
-});
-
 tasks.clean(gulp, {
   dest: './dist/temp'
+});
+
+tasks.copy(gulp, {
+  src: './fixtures/static/*',
+  dest: './dist'
 });
 
 tasks.css(gulp, {
